refactor photo detail module path
