+++ app/build.gradle.kts
@@ -14,7 +14,7 @@ dependencies {
     implementation(project(":feature:home"))
     implementation(project(":feature:collection"))
     implementation(project(":feature:user"))
-    implementation(project(":feature-photo-detail"))
+    implementation(project(":feature:photo-detail"))
     implementation(project(":feature-image-raw"))
     implementation(project(":feature:topics"))
     implementation(project(":feature:search"))
