+++ app/src/main/java/com/stslex/csplashscreen/SplashApplication.kt
@@ -10,7 +10,7 @@ import com.stslex.csplashscreen.core.network.di.moduleCoreNetwork
 import com.stslex.csplashscreen.core.photos.di.moduleCorePhotos
 import st.slex.feature_image_raw.di.moduleRawImage
 import com.stslex.csplashscreen.feature.home.di.moduleFeatureHome
-import st.slex.feature_photo_detail.di.moduleFeaturePhoto
+import com.stslex.csplashscreen.feature.feature_photo_detail.di.moduleFeaturePhoto
 import com.stslex.csplashscreen.feature.search.di.moduleFeatureSearchPhotos
 import com.stslex.csplashscreen.feature.topics.di.moduleFeatureTopics
 import com.stslex.csplashscreen.feature.user.di.moduleFeatureUser
