--- feature-photo-detail/src/main/java/st/slex/feature_photo_detail/data/PhotoRepository.kt
@@ -1,8 +1,0 @@
-package st.slex.feature_photo_detail.data
-
-import kotlinx.coroutines.flow.Flow
-import com.stslex.csplashscreen.core.network.model.ui.ImageModel
-
-interface PhotoRepository {
-    fun getPhotoById(id: String): Flow<ImageModel>
-}
