--- feature-photo-detail/src/main/java/st/slex/feature_photo_detail/ui/components/DetailsImageSetup.kt
@@ -1,8 +1,0 @@
-package st.slex.feature_photo_detail.ui.components
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun DetailImageSetup(){
-
-}
