--- feature-photo-detail/src/main/java/st/slex/feature_photo_detail/ui/utils/DownloadImageUseCase.kt
@@ -1,5 +1,0 @@
-package st.slex.feature_photo_detail.ui.utils
-
-fun interface DownloadImageUseCase {
-    operator fun invoke(url: String, fileName: String)
-}
