+++ feature/photo-detail/.gitignore
