+++ feature/photo-detail/proguard-rules.pro
