+++ feature/photo-detail/src/main/AndroidManifest.xml
