+++ feature/photo-detail/src/main/java/com/stslex/csplashscreen/feature/feature_photo_detail/data/PhotoRepository.kt
@@ -1,0 +1,8 @@
+package com.stslex.csplashscreen.feature.feature_photo_detail.data
+
+import kotlinx.coroutines.flow.Flow
+import com.stslex.csplashscreen.core.network.model.ui.ImageModel
+
+interface PhotoRepository {
+    fun getPhotoById(id: String): Flow<ImageModel>
+}
