+++ feature/photo-detail/src/main/java/com/stslex/csplashscreen/feature/feature_photo_detail/ui/components/DetailsImageSetup.kt
@@ -1,0 +1,8 @@
+package com.stslex.csplashscreen.feature.feature_photo_detail.ui.components
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun DetailImageSetup(){
+
+}
