+++ feature/photo-detail/src/main/java/com/stslex/csplashscreen/feature/feature_photo_detail/ui/utils/DownloadImageUseCase.kt
@@ -1,0 +1,5 @@
+package com.stslex.csplashscreen.feature.feature_photo_detail.ui.utils
+
+fun interface DownloadImageUseCase {
+    operator fun invoke(url: String, fileName: String)
+}
