+++ feature/photo-detail/src/main/res/drawable/ic_baseline_wallpaper.xml
