+++ feature/photo-detail/src/main/res/values/strings.xml
