+++ settings.gradle.kts
@@ -30,7 +30,7 @@ include(":core:collection")
 include(":feature:home")
 include(":feature:collection")
 include(":feature:user")
-include(":feature-photo-detail")
+include(":feature:photo-detail")
 include(":feature-image-raw")
 include(":feature:topics")
 include(":feature:search")
